chore: move deferred
