+++ packages/message-queue/tsconfig.json
@@ -3,8 +3,5 @@
 	"compilerOptions": {
 		"outDir": "dist"
 	},
-	"include": [
-		"src/**/*.ts",
-		"tests/**/*.ts"
-	]
+	"include": ["src/**/*.ts", "tests/**/*.ts"]
 }
